IAR - all within one file
